+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2024-12-18
+
+### Added
+
+- Text over TCP: Added support for C++ and hex style control characters to e.g. add a new line, tab or a carriage return
+  - C++ style characters can be escaped with a single additional backslash (e.g. \\\n)
+  - Hex style characters can be escaped with "0\\\\\\" (e.g. 0\\\\\\0x09)
+
 ## [0.5.0] - 2024-12-15
 
 ### Added